add endpoint to retrieve token icons
