+++ deployables/app/app/tokenList/index.ts
@@ -1,0 +1,2 @@
+export { getAssets } from './getAssets'
+export { getVerifiedAssets } from './getVerifiedAssets'
